Refactored reading of JSON files
